+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - there now is a software repository for **Debian Linux 10.0**
     - please visit the [Debian installation page](https://www.qownnotes.org/installation#Debian)
       for instructions
+- disabled shortcuts for the *Note folders* menu in the *Shortcuts settings*
+  because shortcuts for them are not supported (for [#1235](https://github.com/pbek/QOwnNotes/issues/1235))
 
 ## 19.7.2
 - added a **new editor color schema** *MountainDebo* by @Deboart